Fin de seccion 13

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServer } from '@/auth/actions/auth-actions';
 import prisma from '@/lib/prisma'
 import { NextResponse, NextRequest } from 'next/server'
 import * as yup from 'yup';
@@ -28,10 +29,14 @@ const PostSchema = yup.object({
 })
 
 export async function POST(request: Request) {
+    const user = await getUserSessionServer();
+
+    if (!user) return NextResponse.json('NO Autorizado', { status: 401 });
+
     try {
         const { complete, description } = await PostSchema.validate(await request.json());
 
-        const todo = await prisma.todo.create({ data: { complete, description } });
+        const todo = await prisma.todo.create({ data: { complete, description, userId: user.id as string } });
 
         return NextResponse.json(todo)
 
@@ -41,9 +46,12 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(request: Request) {
+    const user = await getUserSessionServer();
+
+    if (!user) return NextResponse.json('NO Autorizado', { status: 401 });
     try {
-        
-        await prisma.todo.deleteMany({ where: { complete:true} });
+
+        await prisma.todo.deleteMany({ where: { complete: true,userId:user.id } });
 
         return NextResponse.json('Exito')
